requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Single-cell/spatial analysis
scanpy>=1.9.0
squidpy>=1.3.0
anndata>=0.10.0
scvi-tools>=1.0.0
tangram-sc>=1.0.0

# Utilities
tqdm>=4.65.0
dask>=2023.0.0
Pillow>=10.0.0
tifffile>=2023.0.0

# Note: sthd may be a custom package - install separately if needed
# pip install sthd  # or however it's installed

